Revert "add interface and combine in one compoenet"
This reverts commit aa06a00316482cc73a3b62b09da96267385097ee.
